+++ CHANGELOG.rst
@@ -12,7 +12,8 @@ django-fsm-2 4.2.4 2026-03-16
 
 - Fix GET_STATE and RETURN_VALUE without allowed_states defined
 - Introduce ``ANY_STATE`` and ``ANY_OTHER_STATE`` constants
-- Add ``failed_condition`` attribute to ``TransitionNotAllowed`` exception
+- Add ``TransitionConditionsUnmet`` with ``failed_condition`` for unmet conditions
+- Add ``NoTransition`` and ``InvalidTransition`` subclasses for ``TransitionNotAllowed``
 
 
 django-fsm-2 4.2.3 2026-03-15